Persist timezone selection with incident.
Previously, the incident's timezone was stored only in the user's
`localStorage` cache. As such, when the Incident Commander URL was passed to
another teammate, the new incident commander would inadvertently switch the
incident timezone to their local value. To keep the timezone selection
consistent, the `timezoneID` is now being persisted in Firebase.

**NOTE**: The Timezone items didn't have a good entity identifier previously.
To facilitate stable cross-system synchronization, I generated 199 UUIDs and
applied one to each timezone.

Closes #6