Add ability to sort the timeline of updates.
Previously, the timeline was always sorted in an ASC (ascending) direction.
However, this required you to scroll all the way to the bottom of the list in
order to edit the update that you just entered. This commit adds the ability to
sort the timeline in both an ASC and a DESC direction, with the default being
the descending direction. This way, the most recent update is close at hand.

Closes #10